fixed the security in my sql

+++ admin/user_flag.php
@@ -79,14 +79,14 @@
 		$user_flag = $_POST['user_flag'];
 		$userid = $_POST['userid'];
 				
-		if(isset($_POST['add']))
+		if(isset($_POST['add']) && is_numeric($user_flag) && is_numeric($userid))
 		{
 			$query = 'INSERT INTO user_flags (user , flag) VALUES (' . $userid . ',' . $user_flag . ')';
 			$result = mysql_query($query) or die(report_sql_error($query));
 			$output .= 'Flaggan är inlagd!';
 		}
 		
-		elseif(isset($_POST['remove']))
+		elseif(isset($_POST['remove']) && is_numeric($user_flag) && is_numeric($userid))
 		{
 			$query = 'DELETE FROM user_flags WHERE user =' . $userid . ' AND flag =' . $user_flag . ' LIMIT 1';
 			$result = mysql_query($query) or die(report_sql_error($query));
